Analysis: Suppress animation set start and recycle related lint warnings
Warning Message: "This animation should be started with #start()"

Explanation: "Many resources, such as TypedArrays, VelocityTrackers,
etc., should be recycled (with a recycle() call) after use. This lint
check looks for missing recycle() calls."

------------------------------------------------------------------------

These Lint warnings are suppressed, that is, instead of them being
resolved, since a resolution would require a proper investigation and
testing. As such, it might be best to ignore this as out of scope, for
now, and so as to not introduce any breaking changes to this animation
related functionalities that are related to 'AnimatorSet'.

PS: Also the fact that the warning mentions 'start()' not being
called, while it is, but it is calling upon missing 'recycler()' call,
is a bit confusing.