Fix: Resolve live data utils unresolved reference transformations error
Warning Message: "Unresolved reference: Transformations"

Explanation: "'Transformations' is now written in Kotlin. This is a
source incompatible change for those classes written in Kotlin that were
directly using syntax such as 'Transformations.map' - Kotlin code must
now use the Kotlin extension method syntax that was previously only
available when using 'lifecycle-livedata-ktx'. When using the Java
programming language, the versions of these methods that take an
'androidx.arch.core.util.Function' method are deprecated and replaced
with the versions that take a Kotlin 'Function1'. This change maintains
binary compatibility."

For more info see 'lifecycle#2.6.0-alpha05' release:
https://developer.android.com/jetpack/androidx/releases/
lifecycle#2.6.0-alpha05

------------------------------------------------------------------------

Simply replacing 'Transformations.map(liveData) { ... }' with the
corresponding 'liveData.map { ... }' extension function resolves this
'LiveDataUtils.kt' specific error.

You will also notice the extra double exclamation ('!!') added on the
'mapNullable(...)' extension function for 'LiveData'. This is done
because otherwise there is a type mismatch between the nullable and
non-nullable returned type of 'LiveData', which would otherwise require
a signature change to properly resolve.

As such, this is a temporary patch, in order to be able to continue with
this 'LifeCycle' update of '2.6.1' version. Later on, and as a separate
task, this whole 'LiveDataUtils.kt' utility class could be revisited to
understand it more, it terms of which of those utility related extension
functions are worth keeping and which could be simply replaced with the
what is officially supported and being offered already via the
'LifeCycle LiveData' and 'LifeCycle LiveData KTX' official libraries.