Analysis: Suppress launch when started lifecycle scope warnings
Warning Message: "'launchWhenStarted(suspend CoroutineScope.() -> Unit):
Job' is deprecated. launchWhenStarted is deprecated as it can lead to
wasted resources in some cases. Replace with suspending
repeatOnLifecycle to run the block whenever the Lifecycle state is at
least Lifecycle.State.STARTED."

Explanation: https://issuetracker.google.com/issues/270049505#comment2

For more info see 'lifecycle#2.6.0-alpha04' release:
https://developer.android.com/jetpack/androidx/releases/
lifecycle#2.6.0-alpha04

------------------------------------------------------------------------

These warnings are suppressed, that is, instead of them being resolved,
since a resolution would require a proper investigation and testing. As
such, it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this scroll related
functionality overall.

For someone to resolve those, as per the explanation link comment above,
one would need to pick one of the below three options:
1) Let the suspending code continue to run to completion.
2) Cancel the suspending code and restart it when you come back above
that state.
3) Cancel the suspending code and don't restart it.