Fix: Resolve parameter specified as non-null is null crashes
Crash Message: "java.lang.NullPointerException: Parameter specified as
non-null is null: method org.wordpress.android.*.Xyz$setupObservers
$X.invoke, parameter it"

Explanation: "The nullability of the 'onChanged' method of a Observer
created in Kotlin now matches the nullability of the generic type. If
you want 'Observer.onChanged()' to accept a nullable type, you must
instantiate the 'Observer' with a nullable type."

For more info see 'lifecycle#2.6.0' release:
https://developer.android.com/jetpack/androidx/releases/lifecycle#2.6.0

Below is one example of such crash on 'ActivityLogListFragment'
(out of many) for when the user is trying to enter the activity
log screen.

------------------------------------------------------------------------
E  FATAL EXCEPTION: main
   Process: com.jetpack.android.beta, PID: 31318
   java.lang.NullPointerException: Parameter specified as non-null is
    null: method org.wordpress.android.ui.activitylog.list
    .ActivityLogListFragment$setupObservers$9.invoke, parameter it
     at org.wordpress.android.ui.activitylog.list
      .ActivityLogListFragment$setupObservers$9.invoke(Unknown Source:2)
     at org.wordpress.android.ui.activitylog.list
      .ActivityLogListFragment$setupObservers$9.invoke
      (ActivityLogListFragment.kt:213)
     at org.wordpress.android.ui.activitylog.list
      .ActivityLogListFragmentKt$sam$androidx_lifecycle_Observer$0
      .onChanged(Unknown Source:2)
     at org.wordpress.android.viewmodel.SingleLiveEvent$observe$1
      .onChanged(SingleLiveEvent.kt:50)
------------------------------------------------------------------------

As per the explanation above, it seems that the crash mainly happens on
'SingleLiveEvent<Unit>' type of 'LiveData' being observed, and that,
because of the fact that there exist this 'observer.onChanged(t)' call
within this overridden 'SingleLiveEvent.observe(...)' function. As such,
all those 'SingleLiveEvent<Unit>' got updated to their nullable 'Unit'
equivalent, the 'SingleLiveEvent<Unit?>', which fixes any issues.

FYI: This 'SingleLiveEvent' helper class needs to be replaced with a
more modern approach in order to avoid any future such surprises.