Deps: Add androidx core transitively used dependencies on image editor
It is generally recommended that transitively used dependencies should
be declared directly.

FYI: Actually, if that is not done, the 'image-editor' module would
still be using the '1.8.0' version of 'androidx.core:core' instead of
the globally defined '1.10.0' version. And this, actually shows the
danger with transitively used dependencies that are not declared
directly.