Analysis: Resolve composable naming lint warning on loaded state
Warning Message: "Composable functions with a return type should start
with a lowercase letter"

Explanation: "@Composable functions without a return type should use
similar naming to classes, starting with an uppercase letter and ending
with a noun. @Composable functions with a return type should be treated
as normal Kotlin functions, starting with a lowercase letter."

------------------------------------------------------------------------

Renaming the 'LoadedState(...)' composable function, that has a return
type, to the lowercase letter 'loadedState(...)' fixes this warning.