Refactor: Change live data utils map nullable mapper lambda signature
This change removes the need to have this double exclamation ('!!')
added on the 'mapNullable(...)' extension function for 'LiveData', but
that change updates this extension function signature nonetheless.

However, the client classes that are using this 'mapNullable(...)'
extension function doesn't seem to be affected in any way, which makes
this change look safe, at least as far as it was tested, via unit, UI
and manual testing. For completeness and visibility purposes, below are
those client classes:
1) [StatsViewAllViewModel] Via the 'dateSelectorData' field.
2) [StatsViewModel] Via the 'toolbarHasShadow' field.
3) [StatsListViewModel] Via the 'dateSelectorData' field.
4) [WPMainActivityViewModel] Via the 'onFocusPointVisibilityChange'
field.