Analysis: Suppress restricted api lint error on wp dialog snackbar
Error Message: "SnackbarBaseLayout.setLayoutParams can only be called
from within the same library group (referenced groupId=com.google
.android.material from groupId=WPAndroid)"

Explanation: "This API has been flagged with a restriction that has not
been met.

Examples of API restrictions:
* Method can only be invoked by a subclass
* Method can only be accessed from within the same library (defined by
the Gradle library group id)
* Method can only be accessed from tests.

You can add your own API restrictions with the @RestrictTo annotation."

------------------------------------------------------------------------

This Lint error is suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this update overall.