Tool: Remove google material dependency from ignored dependabot updates
After updating the 'Material Component' library to its latest '1.9.0'
version, this ignore configuration is not longer relevant for this
library.

FYI: Actually, this ignore configuration wasn't relevant for this
library for some time now. This is because the 'Material Component'
library was anyway pointing to '1.6.0-alpha01' for some time now, and
that, due to transitive updates.