Deps: Add firebase iid transitively used dependency on wordpress
It is generally recommended that transitively used dependencies should
be declared directly.

------------------------------------------------------------------------

FYI: In this specific case and as per the Firebase 'Remote Config'
version '19.2.0; release note, when using this library, if explicitly
calling 'Instance ID' APIs and relying on an indirect dependency for the
'Instance ID' SDK, the app needs to be updated to either have a direct
dependency on the 'Instance ID' SDK added, or to use the new Firebase
installations SDK to access client identifiers, which include Firebase
installation IDs or installation auth tokens.

For more info see: https://firebase.google.com/support/release-notes/
android#remote-config_v19-2-0