Deps: Update firebase config to 20.0.4
Release Notes: https://firebase.google.com/support/release-notes/
android#remote-config_v20-0-3

------------------------------------------------------------------------

FYI: Added support for other Firebase products to integrate with Remote
Config.

PS: It seems that although there exist a release for '20.0.4', the
release notes for '20.0.4' are not available. Thus, the '20.0.4' update,
but at the same time, having the link for the '20.0.3' release notes.