Deps: Add firebase iid transitively used dependency on wordpress
It is generally recommended that transitively used dependencies should
be declared directly.

------------------------------------------------------------------------

FYI: In this specific case and as per the Firebase 'Messaging'
version '22.0.0' release notes, there exist a caution that this is a
breaking change for apps that use FCM and the deprecated Firebase
Instance ID API to manage registration tokens.

First and foremost, it is strongly recommended to migrate to FCM's
token APIs. But, if that is not possible, then, adding a direct
dependency on the 'firebase-iid' library is the recommended way going
forward, which is exactly what this commit is about.

For more info see: https://firebase.google.com/support/release-notes/
android#messaging_v22-0-0