Deps: Update firebase messaging to 23.0.1
Release Notes: https://firebase.google.com/support/release-notes/
android#messaging_v23-0-1

------------------------------------------------------------------------

FYI: On Android 7.0 and earlier, the SDK now logs that a notification
was opened after onActivityCreated to avoid a race condition when
unparceling the extras Bundle.

PS: Note the fact that this update is actually failing to build on 2
files found with path 'META-INF/DEPENDENCIES' from inputs
(due to the 'httpcomponents' transitive dependency). Updating to
'23.0.2' fixes this build failure.