fix Blob URLs opening externally in Site Editor
Blob URLs (starting with "blob:") are browser generated URLs for blob objects:
https://developer.mozilla.org/en-US/docs/Web/API/Blob

Those URLs are specific to the browser instance that generates them so they
can't and should not be opened externally.

It looks like there were updates to the Site Editor page, which we open in our
WebView, that introduced the usage of such URLs, but our client is intercepting
them and trying to open them externally instead of just loading the resources,
causing the page to not load properly.