remove setRetainInstance from AppSettingsFragment
It doesn't seem it is needed and retaining the state causes the fragment to not
be recreated (with the correct UI colors) when the night mode is set with the
screen open.

I couldn't find any reason we need to keep the fragment retained and things seem
to work fine without it.