remove retainInstance from AccountSettingsFragment
It doesn't seem it is needed and retaining the state causes the fragment to not
be recreated (with the correct UI colors) when the night mode is set with the
screen open.

Looking at the codebase history it seems the fragment was retained purely to
keep the dialogs open on screen rotation instead of handling save/restore state.
For screen rotation this is not neede anymore, but we need to properly handle
the custom dialog preferences state save/restore for this screen to be fully
working correctly now:
- EditTextPreferenceWithValidation
- SummaryEditTextPreference