dismiss the dialog on destroy to avoid window leaks
Removing `setRetainInstance` caused the DetailListPreference to leak the Dialog
window when the Activity is destroyed due to system initiated destruction and
re-creation, because the old Dialog was not dismissed. To prevent this, we now
dismiss the dialog inside `onSaveInstanceState`. This is not the best place to
do it since it's not an actual lifecycle event, but it's a pretty good clue that
the preference is being destroyed due to configuration changes (such as system
dark mode toggling). Also, we will likely need to rewrite these preferences when
moving to either androidx.preference or Compose, so this fix should work fine in
the meantime.