add save/restore state handling in SummaryEditTextPreference
This introduces changes similar to what was done in DetailListPreference, by
implementing the Save/Restore instance state methods to deal with the custom
dialog spawned by this preference.

This makes sure that in system initiated process destroy/re-creation, the dialog
will be shown properly on screen without losing non-persisted data, which makes
it work in Preference Fragments that are not retained during system dark theme
toggling, for instance.