let the text be restored in EditTextPreferenceWithValidation
When `showDialog` is called with non-null bundle it means that the text will be
restored via the parent's `onRestoreInstanceState` method. This class was
overwriting that text with the summary logic (which sets the initial text). The
text only needs to be initialized from the summary if the bundle is null.