Build: Enable non transitive resources for the project
This commit updates the encrypted 'gradle.properties.enc' file to point
to that pinned hash which has the 'android.nonTransitiveRClass=true'
property added into the 'gradle.properties' file for the WPAndroid
project.

FYI: This is done in order for the CI to pick-up this specific change
and download the correct version of 'gradle.properties' to then build
the project with non transitive resources enabled.