Update PrepublishingHomeSocialItem to match prepublish items
This makes sure this item looks visually like the other items used in the same
bottom sheet, which come from XML and were not using the same text styles.

This also hoists the background color so the caller can set it and it is
correctly passed to the TrainOfAvatars border color.