Analysis: Suppress internal inset res lint warning for media settings
Warning Message: "Using internal inset dimension resource
status_bar_height is not supported"

Explanation: "The internal inset dimension resources are not a supported
way to retrieve the relevant insets for your application. The insets are
dynamic values that can change while your app is visible, and your app's
window may not intersect with the system UI. To get the relevant value
for your app and listen to updates, use
androidx.core.view.WindowInsetsCompat and related APIs."

------------------------------------------------------------------------

This Lint warning is suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.