Analysis: Suppress discouraged api lint warning for media settings
Warning Message: "Use of this function is discouraged because resource
reflection makes it harder to perform build optimizations and
compile-time verification of code. It is much more efficient to retrieve
resources by identifier (e.g. R.foo.bar) than by name
(e.g. getIdentifier("bar", "foo", null))."

Explanation: "Discouraged APIs are allowed and are not deprecated, but
they may be unfit for common use (e.g. due to slow performance or subtle
behavior)."

------------------------------------------------------------------------

This Lint warning is suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.