Analysis: Resolve kotlin nullness annotation lint warnings
Warning Message: "Do not use @NonNull in Kotlin; the nullability is
already implied by the Kotlin type ReaderPost not ending with ?"

Explanation: "In Kotlin, nullness is part of the type system; s: String
is never null and s: String? is sometimes null, whether or not you add
in additional annotations stating @NonNull or @Nullable. These are
likely copy/paste mistakes, and are misleading."