Analysis: Suppress custom splash screen lint warning for wp launch
Warning Message: "The application should not provide its own launch
screen"

Explanation: "Starting in Android 12 (API 31+), the application's Launch
Screen is provided by the system and the application should not create
its own, otherwise the user will see two splashscreens. Please check the
SplashScreen class to check how the Splash Screen can be controlled and
customized."

More info: https://developer.android.com/guide/topics/ui/splash-screen

------------------------------------------------------------------------

This Lint warning is suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.