Analysis: Suppress non constant resource id lint warnings
Warning Message: "Resource IDs will be non-final by default in Android
Gradle Plugin version 8.0, avoid using them in const fields"

Explanation: "Avoid the usage of resource IDs where constant expressions
are required.

A future version of the Android Gradle Plugin will generate R classes
with non-constant IDs in order to improve the performance of incremental
compilation."

------------------------------------------------------------------------

This Lint warnings are suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.

FYI: This change relates to 20d305ced43c1a9cca3adb12e26a27643cdf176e
and the fact that non final resources has been disabled for the project.
Having this Lint warnings activated is great as it will guard again that
in the future by not allowing any more such usages of non final
resources throughout the project.