Analysis: Suppress visible for tests id lint warning
Warning Message: "This method should only be accessed from tests or
within private scope"

Explanation: "With the @VisibleForTesting annotation you can specify an
otherwise= attribute which specifies the intended visibility if the
method had not been made more widely visible for the tests.

This check looks for accesses from production code (e.g. not tests)
where the access would not have been allowed with the intended
production visibility."

------------------------------------------------------------------------

This Lint warning is suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.