Analysis: Resolve obsolete sdk int lint warning for wp web view
Warning Message: "Unnecessary; SDK_INT is never < 24"

Explanation: "This check flags version checks that are not necessary,
because the minSdkVersion (or surrounding known API level) is already at
least as high as the version checked for.

Similarly, it also looks for resources in -vNN folders, such as
values-v14 where the version qualifier is less than or equal to the
minSdkVersion, where the contents should be merged into the best
folder."

------------------------------------------------------------------------

Remove the no longer necessary 'getModifiedContextForOlderAPIs()'
extension function altogether and any StackOverflow reference to this
fix is enough.