Add a condition to prevent double adding a fragment
It isn't expected that we'd need this, based on the documentation of
`SingleLiveEvent`, however, it seems that somehow this event is being
triggered multiple times. While this should fix the crash, it may also
be worth tracking down whether there is an issue with `SingleLiveEvent`
that could be a cause for other similar problems.