Analysis: Suppress invalid color hex value lint warnings
Warning Message: "Invalid Color hex value"

Explanation: "Creating a Color with a hex value requires a 32 bit
value (such as 0xFF000000), with 8 bits being used per channel (ARGB).
Not passing a full 32 bit value will result in channels being
undefined / incorrect."

------------------------------------------------------------------------

These Lint errors are suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.

PS: Those Lint errors were also removed from the Lint baseline, and
suppressed locally and closer to source, as the package path is no
longer valid, and thus, they started to fail in the first place.