Build: Replace dependency substitution substitute with method with using
This is done because otherwise local builds can't be configured due a
Gradle API update that is causing a failures such as the below for the
'FluxC' composite build:

------------------------------------------------------------------------

FAILURE: Build failed with an exception.

* Where:
Script '/Users/.../config/gradle/included_builds.gradle' line: 35

* What went wrong:
No signature of method: org.gradle.api.internal.artifacts.ivyservice
 .dependencysubstitution.DefaultDependencySubstitutions$1.with()
 is applicable for argument types: (org.gradle.internal.component.local
  .model.DefaultProjectComponentSelector)
  values: [project :WordPress-FluxC-Android:fluxc]
Possible solutions: with(groovy.lang.Closure), with(boolean,
 groovy.lang.Closure), wait(), wait(long), find(), wait(long, int)
> No signature of method: org.gradle.api.internal.artifacts.ivyservice
 .dependencysubstitution.DefaultDependencySubstitutions$1.with()
 is applicable for argument types: (org.gradle.internal.component.local
  .model.DefaultProjectComponentSelector)
  values: [project :WordPress-FluxC-Android:fluxc]
  Possible solutions: with(groovy.lang.Closure), with(boolean,
   groovy.lang.Closure), wait(), wait(long), find(), wait(long, int)

------------------------------------------------------------------------

PS: Only 'localAboutAutomatticPath' was already configured to be using
the up-to-date 'using' method, that is, in favor of the 'with', which
verifies that this is indeed the right way forward.