create unit test to emulate crash scenario
This test confirms that having the extra `mutex.unlock` is the root cause of the
crash because race conditions on ThreadPools with multiple threads can cause
that line to be called while Mutex internals are dealing with coroutine
cancellation, which also calls unlock internally, leading to `unlock` being
called when the Mutex is not locked, causing the crash.

This also made it clear that we are using the IO dispatcher for doing processing
which is not optimal as well, as the number of threads in the IO dispatcher can
be theoretically infinite, and is at least 64, which greatly increases the
chance for this kind of race conditions.