Make prepublishing sheet scrollable and non-collapsible
Revert the previous Revert, since the fix was incorrectly pushed directly to
the release branch. This adds back the fix.

This is needed especially for Story Posts as the Story post title item is not
part of the RecyclerView, causing it to be pushed out of the Fragment bounds
when the RecyclerView is too big (since this RecyclerView shows all content at
once).