Remove usage of getLastRecordedVideoUri and use data from result
When using the ACTION_VIDEO_CAPTURE intent, the Camera returns the video URI in
the data field of the result Intent in `onActivityResult`. This is the preferred
way of retrieving that URI (when EXTRA_OUTPUT is not set in the request intent)
because it already provides the saved file URI with access permissions granted
to the caller application.

This is perfect for Android 14 because of the new partial permission grant for
videos and images, which makes getLastRecordedVideoUri to stop working in that
scenario.

The getLastRecordedVideoUri is being removed from the WordPress-Utils lib to
avoid it being used as well.