Stop requesting unnecessary storage permissions
When other apps share or drop (in the case of Aztec) images/videos to our app,
the permission to the media from those shared URIs is granted to us, so there's
no need to request for extra storage permissions.

Requesting READ_MEDIA permissions in Android 14 in those cases is even worse if
user has given only partial media permission, as it's not clear what we are
requesting (and not needed anyway).