Analysis: Make intent non null on the on new intent functions for kotlin
This is done so as to have the 'onNewIntent(...)' function for Kotlin
files match the 'onNewIntent(...)' method for Java files. Otherwise,
half of those files were using 'intent' as nullable and the other half
were using 'intent' as non null, which creates confusion.

An example of existing Kotlin activities that are already using 'intent'
as non-null are:
- PagesActivity.kt
- PostsListActivity.kt

PS: All the Kotlin activities in this change are 'androidx.fragment.app'
related and more specifically inherit from 'FragmentActivity.java'.