Replace guessUrl with guessFileName
Not sure if this feature has been broken all this time or something changes in
how that Android method works, but `guessUrl` actually returns a valid URL,
which most of the time is NOT a valid file name, as it contains several `.`, `:`
and `/` characters, so the DownloadManager is not able to download the files.

`guessFileName` on the other hand transforms URLs to valid file names and should
be used in this scenario.