Analysis: Remove unused lint xml file from fluxc module
It doesn't seem that this 'lint.xml' file is being within the 'fluxc'
module, or anywhere for that matter. Thus, this is now being removed so
as to avoid having to worry about this unused file, thinking that it has
some kind of effect on this project's current Lint configuration.