Analysis: Add missing n-a to taxonomy model
FYI: 'n-a' stands for 'nullability annotations'.

This 'TaxonomyModel' is never being used in production, that is, at
least via the JP/WPAndroid app. As such, the nullability on the 'mName',
'mLabel' and 'mDescription' fields are being dictated by tests. As such,
and because every test that is using a 'TaxonomyModel' is also setting
its 'mName' right after, this is why this field is being annotated as
'@NonNull'. The same does not apply for the 'mLabel' and 'mDescription'
fields and as such, those field are being annotated as '@Nullable'.

PS: This 'NotNullFieldNotInitialized' warning got suppressed because a
table related 'model' class can never have its fields initialized via a
constructor initialization, or otherwise for that matter.