Analysis: Add missing n-a to fetch terms response payload
FYI: 'n-a' stands for 'nullability annotations'.

PS.1: This 'NotNullFieldNotInitialized' warning got suppressed because a
'payload' static class has both, a 'normal' and an 'error' payload.
Although a 'normal' payload could have its fields annotated with
'@NonNull', that is, when they are being instantiated as such via its
'normal' constructor, because of the 'error' payload, and its
constructor not needing most of those necessarily, this '@NonNull'
condition might no longer be valid for both those 'payload' classes.
PS.2: With this change, and because an 'error' payload needs to always
go via an 'payload.isError()' check, it is assumed that those
'@NonNull' fields will never be accessed, that is when the
'payload.isError()' check returns true. Thus, and in order to have a
better nullability throughput when on the 'normal' branch, it has been
decided in this change to suppress this 'NotNullFieldNotInitialized'
warning and ignore cases where those fields are incorrectly access, when
on the 'error' branch, which, unfortunately, could potentially cause a
NPE. To make this a bit more clear, an extra comment is added on those
fields that are '@NonNull' on both, the 'normal' and 'error' branches.
This is done in order for future authors to understand which field(s)
can be utilized when on the 'error' branch, and which can't/shouldn't.