Analysis: Resolve unnecessary safe call and condition compile warnings
Compile Warnings:
- "Unnecessary safe call on a non-null receiver of type TermModel"
- "Condition 'event.term.name == null' is always 'false'"

------------------------------------------------------------------------

As per the removed comment, and although it specifies that sometimes the
API will return a success response with a null name, which then will be
treated as an error, because without a name, there is no category, after
some testing, it has been verified that this is not needed anymore. A
'409 duplicate' error is returned instead, that is, instead of a valid
response but with a 'null' name, which the deleted code was guarding
against.

This previous 'term.name == null' change is related to this JP/WPAndroid
#13256 PR below:

https://github.com/wordpress-mobile/WordPress-Android/pull/13256

However, as per this newly merge FluxC #2826 PR below, the
'TermWPComRestResponse' response class and its 'name' field was marked
as '@NonNull'.

https://github.com/wordpress-mobile/WordPress-FluxC-Android/pull/2826

PS: If and whenever the backend starts sending a 'term.name' as
'nullable', then this need to be addressed on the backend side, that is,
instead of making 'term.name' a nullable field on the FluxC side and
start guarding against such occurrences everywhere.