Analysis: Make menu nn-a on the on create options menu fun for kotlin
FYI: 'nl-a' stands for 'non-null annotation'.

This is done so as to have the 'onCreateOptionsMenu(...)' function for
Kotlin files match the 'onCreateOptionsMenu(...)' method for Java files.
Otherwise, one of those files were using 'menu' as nullable and the rest
of them were using 'menu' as non-null, which creates confusion.

An example of existing Kotlin activities that are already using 'menu'
as non-null are:
- ScanActivity.kt
- PostsListActivity.kt

PS: The Kotlin activity in this change is 'android.app' related and more
specifically inherit from 'Activity.java'.