Analysis: Suppress non constant res id lint warning for activities
Warning Message: "Resource IDs will be non-final by default in Android
Gradle Plugin version 8.0, avoid using them in const fields"

Explanation: "Avoid the usage of resource IDs where constant expressions
are required.

A future version of the Android Gradle Plugin will generate R classes
with non-constant IDs in order to improve the performance of incremental
compilation."

------------------------------------------------------------------------

This Lint warning is suppressed, that is, instead of it being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.

Related Commit: f8bc363cf39c000216b53bc8e910b00f5b6e1a89