Analysis: Make mode & menu nn-a on the on prp action mode fun for kotlin
FYI: 'nl-a' stands for 'non-null annotation'.

This is done so as to have the 'onPrepareActionMode(...)' function for
Kotlin files match the 'onPrepareActionMode(...)' method for Java files.

Although all existing Kotlin files were using 'mode' & 'menu' as
nullable, none of them are actually using either 'mode' or 'menu' at
all. Thus, this change updates those files to match the nullability
of 'mode' & 'menu' on the 'onCreateActionMode(...)' and
'onActionItemClicked(...)', where those are being utilized as non-null.