Refactor: Use non-null viewbinding method parameter for site picker
This is done in order to avoid using the 'if (mBinding != null) { ... }'
check everywhere and only use that where absolutely necessary, which
usually is only needed when a 'binding' is required via an SDK related
'@Override' method, that is, it being the 'binding' starting point.

FYI: This way, when this class gets converted to Kotlin (in the future),
instead of using the 'binding' as a parameter to a method, the same
method will get converted to a 'binding' extension function, and then,
everything will become much more straightforward. For what is worth,
this pattern is already used as such with Kotlin classes. For example,
see 'PreviewImageFragment.kt' class and its 'initializeViews()'
extension function.