Refactor: Replace findviewbyid with viewbinding for comment detail (1/2)
This change includes the inner 'mReplyBinding' binding, which is an
included layout within the 'comment_detail_fragment.xml' file, with the
'layout_comment_box' id, which points to the
'reader_include_comment_box.xml' layout file.

FYI: In order to reference the included 'layout_comment_box' id, the
'layout_container' id needed to be created on the top level view of the
the 'reader_include_comment_box.xml' layout file.