Refactor: Replace findviewbyid with viewbinding for comment detail (2/2)
This change includes the outer 'mActionBinding' binding, which is added
to the main binding's 'commentContentContainer' layout programmatically,
which points to the 'comment_action_footer.xml' layout file.

FYI: The 'layout_buttons' id is already created on the top level view
of the the 'comment_action_footer.xml' layout file. As such, referencing
this is already possible.