Refactor: Use non-null viewbinding method param for comment detail (1/3)
This is done in order to avoid using the 'if (mBinding != null) { ... }'
check everywhere and only use that where absolutely necessary, which
usually is only needed when a 'binding' is required via an SDK related
'@Override' method, that is, it being the 'binding' starting point.

FYI: This way, when this class gets converted to Kotlin (in the future),
instead of using the 'binding' as a parameter to a method, the same
method will get converted to a 'binding' extension function, and then,
everything will become much more straightforward. For what is worth,
this pattern is already used as such with Kotlin classes. For example,
see 'PreviewImageFragment.kt' class and its 'initializeViews()'
extension function.

PS: An exception to this were the 'setComment(...)' methods, and the
fact that the non-null 'binding' viewbinding wasn't added as a parameter
to those methods. This was done because these methods are also called
from within 'onCreate(...)' SDK related '@Override' fragment lifecycle
method, which is called before the 'onCreateView(...)' SDK related
'@Override' fragment lifecycle method. As such, and because the
'mBinding' field is only set during the 'onCreateView(...)' SDK related
'@Override' fragment lifecycle method, it would always be null during
the 'onCreate(...)' SDK related '@Override' fragment lifecycle method,
causing a logical initialization change on this comment detail screen,
which will most probably cause an issue due to the fact that
the 'mComment', 'mSite' and 'mIsUsersBlog' fields would not be set,
something that is actually done via the 'setComment(...)' method itself
and before it needing to check and access a non-null 'mBinding'. Same
logical thinking and process was applied to 'setNote(...)' and the fact
that the 'mNote' and 'mSite' fields would not be set otherwise.