Refactor: Remove unnecessary is added check for show comment
The 'isAdded()' check is already available within the 'showComment(...)'
method itself, on the very first statement of the method, which returns
immediately if 'isAdded()' is not met. As such, having any client that
is calling 'showComment(...)' check that again is unnecessary, plus,
this unnecessarily increases the complexity of the class overall.

FYI: Actually, the 'onStart(...)' SDK related '@Override' fragment
lifecycle method is already depending on this fact.