Refactor: Use non-null viewbinding method param for comment detail (2/3)
This is done in order to avoid using the
'if (mReplyBinding != null) { ... }' check everywhere and only use that
where absolutely necessary, which usually is only needed when a
'replyBinding' is required via an SDK related '@Override' method, that
is, it being the 'replyBinding' starting point.

FYI: This way, when this class gets converted to Kotlin (in the future),
instead of using the 'replyBinding' as a parameter to a method, the
same method will get converted to a 'replyBinding' extension function,
and then, everything will become much more straightforward. For what is
worth, this pattern is already used as such with Kotlin classes. For
example, see 'PreviewImageFragment.kt' class and its 'initializeViews()'
extension function.

PS: As per the previous commit message description, an expeption to this
were the 'setComment(...)' and 'setNote(...)' methods, for the very same
reasons.