Analysis: Suppress same parameter value on comment detail
Warning: "Actual value of parameter 'forceLike' is always 'false'"

FYI: Currently 'likeComment(actionBinding, false)' is used once and
'likeComment(actionBinding, true)' is never used, but this doesn't mean
it will not be used in the future, or wasn't used in the past. As such,
keeping this method signature and suppressing this warning for the time
being seems to be the better approach here.