Analysis: Add missing nl-a to comment field
FYI: 'nl-a' stands for 'nullable annotation'.

PS: This change also removes the below dead code situations:
1. editComment(...) method: The 'mComment == null' that guards the
access to 'intent.putExtra(...)' can never be met. This is due to the
fact that this method was anyway returning on 'mComment' being null.
2. updateStatusViews(...) method: The 'mComment != null' is always true
and thus 'mNote.hasLikedComment()' can never be triggered. Again, this
is due to the fact that this method was anyway returning on 'mComment'
being null.