Analysis: Utilize nl-a on node field
FYI: 'nl-a' stands for 'nullable annotation'.

PS: Contrary to the 'mComment' field, which is usually passed as
'@NonNull' to most methods, the 'mNote' field is usually passes as
'@Nullable' to most methods. This is because a 'note' is only used when
a comment is accessed via the 'Notifications' tab, while a 'comment' is
universal to both, the 'Comments' screen and 'Notifications' tab.