Analysis: Suppress notify data set changed lint warn on reader cmnt list
Warning: "It will always be more efficient to use more specific change
events if you can. Rely on notifyDataSetChanged as a last resort.""

Explanation: "The RecyclerView adapter's onNotifyDataSetChanged method
does not specify what about the data set has changed, forcing any
observers to assume that all existing items and structure may no longer
be valid. LayoutManagers will be forced to fully rebind and relayout all
visible views."

------------------------------------------------------------------------

This Lint warning is suppressed, that is, instead of it being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore this as out of scope, for now, and so as to not
introduce any breaking changes with this upgrade overall.